Add test supabase module
