+++ lib/libraries/supabase/testing/supabase_test_module.dart
@@ -1,0 +1,17 @@
+import 'package:construculator/libraries/config/config_module.dart';
+import 'package:construculator/libraries/supabase/interfaces/supabase_wrapper.dart';
+import 'package:construculator/libraries/supabase/testing/fake_supabase_wrapper.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+
+class SupabaseTestModule extends Module {
+  @override
+  List<Module> get imports => [
+    ConfigModule(),
+  ];
+  @override
+  void exportedBinds(Injector i){
+    i.addLazySingleton<SupabaseWrapper>(
+      () => FakeSupabaseWrapper(),
+    );
+  }
+}
